Avocado compatibility: introduce utility for setting options
With defaults coming from config file, via the settings module *when
the Avocado implementation of the settings module is concerned only
with the configuration file*. In this case, we can get the value that
will serve as the option defaults from it.

When the Avocado implementation is the "new settings" module, that
concentrates information about the configuration file and command
line, and defaults are registered ahead of time, this becomes a
no-op.

Signed-off-by: Cleber Rosa <crosa@redhat.com>

PoC for setting option with default from config file [2]

Signed-off-by: Cleber Rosa <crosa@redhat.com>