Fix selftests.unit.test_cartesian_config.CartesianConfigTest.testHugeTest1
On commit 027c9682b the cpu_chk_cmd was changed, but the recorded
result, used for comparison, wasn't.

Then, on commit 50bc6b87, the vcpu_thread_pattern assignment was
removed from the input, but not from the result representation.

Finally, on commit a024e512, "extra_params" was changed, but again,
not reflected on the result representation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>