Add support for saving the cartesian config
On most Avocado-VT runs, the cartesian config file will be manipulated
by some command line options.  It's helpful to be able to save the
resulting file so that it can be reused for debugging and
reproducibility reasons.

This saves the configuration after all options command line options,
but *before* the references are processed by the loader code.

As an added note, the resolution of those references use this saved
cartesian config state, and perform "only $REFERENCE" statements to
find only tests matching the references given.  If we were to apply
all those "only $REFERENCE" statements, we'd end up on most cases
with no resulting test.

Signed-off-by: Cleber Rosa <crosa@redhat.com>